test(extraction): organize tests by intent and fix CI marker gaps
Mark the local Transformers extraction tests with ml_vlm so CI runs them
in the dedicated vlm suite instead of the core lane, where torch and
transformers (optional extras they import) are guaranteed present.

Dissolve test_extraction_refinements.py, whose name tracked a refinement
plan rather than intent: move the text-channel page-range cases into
test_extraction_text_channel.py, the DocumentExtractor config and
import-isolation cases into test_extraction.py, and the legacy inline
options adaptation into test_extraction_api.py.

Delete test_granite_vision_extraction.py; its only real coverage
(pipeline routing a local spec to TransformersExtractionModel) moves into
test_extraction_transformers_model.py, dropping the identity-only mock.

Fix the import-isolation test's dependency stubbing: a MetaPathFinder that
raises ModuleNotFoundError replaces sys.modules[name] = None, which left a
None that scipy's torch detection choked on under Python 3.14.

Consolidate the extraction plans into docs/plans/extraction.md.
